backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
sqlmodel==0.0.22
pydantic==2.10.3
python-dotenv==1.0.1
openai>=1.66.2
httpx==0.28.1
openai-agents==0.0.4